fix(core): widen volume accumulators from i64 to i128 to prevent overflow (#88)
SHIBUSDT (price ~$0.00003) regularly accumulated >92.23B tokens per bar
at larger thresholds, overflowing FixedPoint(i64) and producing negative
volumes, corrupted OFI, and broken VWAP. Widen volume/buy_volume/
sell_volume to i128 (proven pattern from turnover), add Python write
guards, detection script, tier1 validation, regression tests, and
pueue-based post-fix repopulation workflow.

SRED-Type: experimental-development
SRED-Claim: RANGEBAR-CORE

+++ crates/rangebar-core/src/interbar.rs
@@ -411,9 +411,10 @@ impl TradeHistory {
         });
 
         // VWAP
-        let total_volume_fp: i64 = lookback.iter().map(|t| t.volume.0).sum();
+        // Issue #88: i128 sum to prevent overflow on high-token-count symbols
+        let total_volume_fp: i128 = lookback.iter().map(|t| t.volume.0 as i128).sum();
         features.lookback_vwap = Some(if total_volume_fp > 0 {
-            let vwap_raw = total_turnover / (total_volume_fp as i128);
+            let vwap_raw = total_turnover / total_volume_fp;
             FixedPoint(vwap_raw as i64)
         } else {
             FixedPoint(0)
